Release 3.1.0

+++ model/src/main/java/org/cbioportal/model/DataBin.java
@@ -1,6 +1,8 @@
 package org.cbioportal.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+
 import org.cbioportal.model.ClinicalDataCountItem.ClinicalDataType;
 
 public class DataBin implements Serializable {
@@ -8,8 +10,8 @@ public class DataBin implements Serializable {
     private String attributeId;
     private ClinicalDataType clinicalDataType;
     private String specialValue;
-    private Double start;
-    private Double end;
+    private BigDecimal start;
+    private BigDecimal end;
     private Integer count;
 
     public String getAttributeId() {
@@ -36,19 +38,19 @@ public class DataBin implements Serializable {
         this.specialValue = specialValue;
     }
 
-    public Double getStart() {
+    public BigDecimal getStart() {
         return start;
     }
 
-    public void setStart(Double start) {
+    public void setStart(BigDecimal start) {
         this.start = start;
     }
 
-    public Double getEnd() {
+    public BigDecimal getEnd() {
         return end;
     }
 
-    public void setEnd(Double end) {
+    public void setEnd(BigDecimal end) {
         this.end = end;
     }
